Pause Functionality & Text Organization
Pause Button now added to screen that will stop music and game time all at once, it is designed to be very easily accessible so that if we wish to add keyboard functionality we can do so. We would have to reassess the input system though if these changes are to be made.

While the buttons are hooked up, the main menu button has been deactivated pending development on scene management bug.

Text organization consisted of moving known scene specific fonts into a separate folder with the idea being that it would create a repeatable structure for any additional fonts. It is fairly miniscule currently until more scene specific fonts can be attributed to it.